Moved delete records tests to AdminClientIntegrationTest
