+++ docs/upgrade.html
@@ -95,7 +95,7 @@
         timeout behavior for blocking APIs. In particular, a new <code>poll(Duration)</code> API has been added which
         does not block for dynamic partition assignment. The old <code>poll(long)</code> API has been deprecated and
         will be removed in a future version.</li>
-    <li>The implementation of DeleteRecordsCommand makes use of the new Java-based AdminClient.</li>
+    <li>The Scala method <code>AdminClient.deleteRecordsBefore</code> has been removed in favor of <code>deleteRecords</code> in the new Java-based AdminClient.</li>
 </ul>
 
 <h5><a id="upgrade_200_new_protocols" href="#upgrade_200_new_protocols">New Protocol Versions</a></h5>
